FLTK3: removed all warnings (shortcut and keys are unsigned int now. We will need the additional bits to allow unicode shortcuts with arbitrary modifier.
git-svn-id: file:///fltk/svn/branches/branch-3.0@8988 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fl_ask.cxx
@@ -239,7 +239,7 @@ static int innards(const char* fmt, va_list ap,
   if (enableHotspot)
     message_form->hotspot(button[0]);
   if (b0 && fltk3::Widget::label_shortcut(b0))
-    button[0]->shortcut(0);
+    button[0]->shortcut((unsigned)0);
   else
     button[0]->shortcut(fltk3::EscapeKey);
